Update setup-edison.md
There is a much better mac-specific section after this section. It is better if mac people skip to that because it has important things to install that this existing section didn't cover. I got hung up until I saw that mac-specific section after this section. It is important to link to it here.